Fix Bluetooth version example and core
